Refactor tool modules into a dedicated `tools` directory
Restructured various utility modules under a new `tools` directory for better organization and modularity. Updated all relevant imports to reflect the new paths, ensuring functionality remains unchanged.
